requirements.txt
python-telegram-bot[job-queue]==20.*
SQLAlchemy==2.*
Pillow
python-dotenv
